Add unit tests for TreatmentRepository functionality
Added a new `TreatmentRepositoryTests` class in the `HairdresserUnitTests` namespace to test the `TreatmentRepository` class. Introduced necessary `using` directives for dependencies. Implemented `[TestInitialize]` and `[TestCleanup]` methods to configure and clean up an in-memory database for testing. Added a test method `Add_ShouldAddTreatmentSuccessfully` to verify treatments can be added to the repository.